+++ src/objects/jcollection.rs
@@ -16,7 +16,11 @@ use super::Reference;
 #[cfg(doc)]
 use crate::errors::Error;
 
-/// Wrapper for `java.utils.Map.Entry` references. Provides methods to get the key and value.
+/// A `java.util.Collection` wrapper that is tied to a JNI local reference frame.
+///
+/// See the [`JObject`] documentation for more information about reference
+/// wrappers, how to cast them, and local reference frame lifetimes.
+///
 #[repr(transparent)]
 #[derive(Debug, Default)]
 pub struct JCollection<'local>(JObject<'local>);
@@ -96,19 +100,26 @@ impl JCollectionAPI {
     }
 }
 
-impl<'local> JCollection<'local> {
+impl JCollection<'_> {
     /// Creates a [`JCollection`] that wraps the given `raw` [`jobject`]
     ///
     /// # Safety
     ///
-    /// `raw` may be a null pointer. If `raw` is not a null pointer, then:
+    /// - `raw` must be a valid raw JNI local reference (or `null`).
+    /// - `raw` must be an instance of `java.util.Collection`.
+    /// - There must not be any other owning [`Reference`] wrapper for the same reference.
+    /// - The local reference must belong to the current thread and not outlive the
+    ///   JNI stack frame associated with the [Env] `'local` lifetime.
+    pub unsafe fn from_raw<'local>(env: &Env<'local>, raw: jobject) -> JCollection<'local> {
+        JCollection(JObject::from_raw(env, raw))
+    }
+
+    /// Creates a new null reference.
     ///
-    /// * `raw` must be a valid raw JNI local reference.
-    /// * There must not be any other `JObject` representing the same local reference.
-    /// * The lifetime `'local` must not outlive the local reference frame that the local reference
-    ///   was created in.
-    pub const unsafe fn from_raw(raw: jobject) -> Self {
-        Self(JObject::from_raw(raw))
+    /// Null references are always valid and do not belong to a local reference frame. Therefore,
+    /// the returned `JCollection` always has the `'static` lifetime.
+    pub const fn null() -> JCollection<'static> {
+        JCollection(JObject::null())
     }
 
     /// Unwrap to the raw jni type.
@@ -268,7 +279,7 @@ impl<'local> JCollection<'local> {
             let iterator = env
                 .call_method_unchecked(self, api.iterator_method, ReturnType::Object, &[])?
                 .l()?;
-            Ok(JIterator::from_raw(iterator.into_raw()))
+            Ok(JIterator::from_raw(env, iterator.into_raw()))
         }
     }
 }
@@ -295,10 +306,10 @@ unsafe impl Reference for JCollection<'_> {
     }
 
     unsafe fn kind_from_raw<'env>(local_ref: jobject) -> Self::Kind<'env> {
-        JCollection::from_raw(local_ref)
+        JCollection(JObject::kind_from_raw(local_ref))
     }
 
     unsafe fn global_kind_from_raw(global_ref: jobject) -> Self::GlobalKind {
-        JCollection::from_raw(global_ref)
+        JCollection(JObject::global_kind_from_raw(global_ref))
     }
 }