+++ src/objects/jlist.rs
@@ -15,8 +15,11 @@ use crate::{
 
 use std::{borrow::Cow, ops::Deref};
 
-/// Wrapper for `java.utils.List` references. Provides methods to get, add, and
-/// remove elements.
+/// A `java.util.List` wrapper that is tied to a JNI local reference frame.
+///
+/// See the [`JObject`] documentation for more information about reference
+/// wrappers, how to cast them, and local reference frame lifetimes.
+///
 #[repr(transparent)]
 #[derive(Debug, Default)]
 pub struct JList<'local>(JObject<'local>);
@@ -50,7 +53,7 @@ impl<'local> From<JList<'local>> for JObject<'local> {
 impl<'local> From<JList<'local>> for JCollection<'local> {
     fn from(other: JList<'local>) -> JCollection<'local> {
         // SAFETY: Any `java.lang.List` is also a `java.util.Collection`
-        unsafe { JCollection::from_raw(other.into_raw()) }
+        unsafe { JCollection::kind_from_raw(other.into_raw()) }
     }
 }
 
@@ -95,14 +98,21 @@ impl<'local> JList<'local> {
     ///
     /// # Safety
     ///
-    /// `raw` may be a null pointer. If `raw` is not a null pointer, then:
+    /// - `raw` must be a valid raw JNI local reference (or `null`).
+    /// - `raw` must be an instance of `java.util.List`.
+    /// - There must not be any other owning [`Reference`] wrapper for the same reference.
+    /// - The local reference must belong to the current thread and not outlive the
+    ///   JNI stack frame associated with the [Env] `'env_local` lifetime.
+    pub unsafe fn from_raw<'env_local>(env: &Env<'env_local>, raw: jobject) -> JList<'env_local> {
+        JList(JObject::from_raw(env, raw))
+    }
+
+    /// Creates a new null reference.
     ///
-    /// * `raw` must be a valid raw JNI local reference.
-    /// * There must not be any other `JObject` representing the same local reference.
-    /// * The lifetime `'local` must not outlive the local reference frame that the local reference
-    ///   was created in.
-    pub const unsafe fn from_raw(raw: jobject) -> Self {
-        Self(JObject::from_raw(raw))
+    /// Null references are always valid and do not belong to a local reference frame. Therefore,
+    /// the returned `JList` always has the `'static` lifetime.
+    pub const fn null() -> JList<'static> {
+        JList(JObject::null())
     }
 
     /// Unwrap to the raw jni type.
@@ -342,10 +352,10 @@ unsafe impl Reference for JList<'_> {
     }
 
     unsafe fn kind_from_raw<'env>(local_ref: jobject) -> Self::Kind<'env> {
-        JList::from_raw(local_ref)
+        JList(JObject::kind_from_raw(local_ref))
     }
 
     unsafe fn global_kind_from_raw(global_ref: jobject) -> Self::GlobalKind {
-        JList::from_raw(global_ref)
+        JList(JObject::global_kind_from_raw(global_ref))
     }
 }